Now using javassist to create proxies that can be garbage collected, fixed querying by association with proxies and added methods to load proxies to the Session interface and GormStaticApi

+++ redis/src/main/java/org/springframework/datastore/redis/util/RedisTemplate.java
@@ -301,7 +301,7 @@ public class RedisTemplate {
     }
 
     public long incr(final String key) {
-        return (Long)execute(new RedisCallback() {
+        return (Integer)execute(new RedisCallback() {
             public Object doInRedis(RedisClient redis) {
                 return redis.incr(key);
             }
